Fix nested container counting safeguards

+++ src/main/java/com/chenjicheng/chestcountoverlay/overlay/CountedItem.java
@@ -2,5 +2,7 @@ package com.chenjicheng.chestcountoverlay.overlay;
 
 import net.minecraft.item.ItemStack;
 
-public record CountedItem(ItemStack stack, long totalCount, int firstSlot) {
+import java.math.BigInteger;
+
+public record CountedItem(ItemStack stack, BigInteger totalCount, int firstSlot) {
 }
